requirements.txt
streamlit
PyMuPDF
Pillow
langchain
chromadb
openai
python-dotenv
transformers
torch 